==> code/1. Scrapping Usernames.ipynb <==
## 1. Scrapping Usernames

In this notebook, I will be scrapping a list of usernames from MyAnimeList(MAL).  
This will be done using the Jikan unofficial MAL API.  
As there are too many users on MAL, I will only scrape the users that are a part of the largest community club.  
This club is called the "Recommendation Club" and has more than 30,000 users.  

import requests
import pandas as pd

from jikanpy import Jikan
jikan = Jikan()

res = requests.get('https://api.jikan.moe/v3/club/20081/')
res.json()

#at time of writing, the largest club (recommendation club) has 31,839 members or 885 pages of 36 users
# url = "https://api.jikan.moe/v3/club/1/members/1"
club_members = []
for i in range(1,886):
    res = requests.get(f"https://api.jikan.moe/v3/club/20081/members/{i}")
    data = res.json()
    for number in range(36):
        try:
            club_members.append(data['members'][number]['username'])
        except:
            break
    print(f'page {i} done and length is {len(club_members)}')
    # res.json()['members'][0]['username']
len(club_members)

club_members = pd.DataFrame(club_members)

club_members.head()

club_members.to_csv('../data/club_members.csv', index = False)

### Scrapped Data
A total of 28,384 usernames were scrapped.  
Some of the usernames were not scrapped successfully as the club has more than 30,000 users.  
This could have been due to accounts being deleted or there were issues with the request and server.
